Re-implement GitHubProject to use the GitHub API
The existing spec ensures it is implemented identically to
GitHubScraper. To do this I needed to add the deep_format_times method
to manually format Time values. This is because times come out of the
API as strings like this:

2016-12-23T12:17:31Z

The Octokit library (well, actually Sawyer) then parses them into Ruby
Time objects. Rails' JSON processing then converts them to time strings
like this:

2016-12-23T12:17:31.000Z

The two strings are different, but they obviously represent the same
time value.

When we remove GitHubScraper, we can remove the deep_format_times
method, but it is useful to prove equivalence between the two
implementations for now.

I've added VCR to record the responses from the API requests so we
don't have to make real requests on each run.

I've also added the git_hub_token environment variable for
authentication. This isn't strictly necessary, but without
authenticating, we're limited to 60 requests per hour, which is pretty
limiting. With authentication, we get 5,000 requests per hour. The
authentication token doesn't need to have any private permissions
because we are only reading public data.