Fix deprecation warnings
We're using autoprefixer-rails, so the linear-gradient will be
autoprefixed.

There are a bunch of other warnings from
vendor/assets/stylesheets/material-components-web.css, which look like
this:

    DEPRECATION WARNING on line 9590 of /home/turnip/Code/wikihouse.cc/vendor/assets/stylesheets/material-components-web.css:
    Sass 3.6 will change the way CSS variables are parsed. Instead of being parsed as
    normal properties, they will not allow any Sass-specific behavior other than #{}.
    For forwards-compatibility, use #{}:

      --variable: #{rgba(255, 255, 255, 0.7)};

Unfortunately, they're false-positives, because this is a normal CSS
file (which is being processed by Sass in order to minify it and so on).

I can't see a way to get the warning to go away, but we shouldn't
actually need to act on it at all.

Closes #34