Use PushType node rendering
Rather than having our own StaticController, I think it would be good to
migrate to letting PushType handle the rendering. This more closely
matches how PushType is supposed to be used, and will also make it
easier for us to add new elements of CMS functionality (for example, we
might want to make new bits of certain pages editable).